requirements.txt
streamlit
openai-whisper
transformers
torch
sentencepiece
pydub 